separate agreement setting components
